nixos/{system,vps}: move system.autoUpgrade to nixos/vps.nix
It seems you can't use nixos-rebuild to commit the upgraded lock file,
but you can still upgrade the system with --no-write-lock-file flag.
Committing this file automatically doesn't seem a good idea either way,
since this would advance the state of the repository and probably
generate merge conflicts if the update is done in multiple places.

So let's set this in VPS to have unattended upgrades for now and remove
it for the rest. In the future I want to bump flake.lock automatically
using GitHub Actions, so a probably better solution would have to just
do a git pull followed bu a nixos-rebuild.