github: set `overprovision-lvm` on maximize-build-space to true
This action assumes that most of build is done inside the workdir,
however in this case most of the space is used on `/nix/store`. This
wasn't a issue before since it created LVM as sparse files, however now
that this is fixed this will cause issues. Setting `overprovision-lvm`
to true will return to the previous behavior.

We could set `root-reserve-tmp` to a bigger value, however we actually
need a lot of space on `/nix/store` and almost none (except for the
actual project) on workdir, so this shouldn't cause any issues.